+++ docs/releases.rst
@@ -13,6 +13,7 @@ unreleased
 `git master <https://github.com/meejah/txtorcon>`_ *will likely become v0.14.0*
 
  * :class:`txtorcon.interface.IStreamAttacher` handling was missing ``None`` and ``DO_NOT_ATTACH`` cases if a Deferred was returned.
+ * add ``.is_built`` Deferred to :class:`txtorcon.Circuit` that get `callback()`d when the circuit becomes BUILT
 
 
 v0.13.0